feat: moved util functions to utils folder
