+++ app/categories/[id]/page.tsx
@@ -1,7 +1,7 @@
 import { notFound } from "next/navigation"
 import { ProductCard } from "@/components/product-card"
 import { categories } from "@/lib/categories"
-import { products } from "@/lib/products"
+import { products } from "@/utils/products"
 
 export default function CategoryPage({ params }: { params: { id: string } }) {
   const category = categories.find((c) => c.id === params.id)
