+++ supabase/.temp/cli-latest
@@ -1,0 +1,1 @@
+v2.20.12
